requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
editdistance==0.6.2
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
macholib==1.15.2
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
openai==0.27.4
pandas==2.0.1
pip==21.2.4
puzpy==0.2.5
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
setuptools==58.0.4
six==1.15.0
sqlitedict==2.1.0
thefuzz==0.19.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.37.0
yarl==1.9.1
zipp==3.15.0
